requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
numpy==2.2.5
pillow==11.3.0
torch==2.10.0
torchvision==0.25.0
scikit-learn==1.7.2
mlflow==3.8.1
matplotlib==3.10.7
pytest==8.3.3
httpx==0.27.2
